chore: move `Frame` to `Frame.ts` (#8806)
