+++ src/common/HTTPRequest.ts
@@ -18,7 +18,7 @@ import {ProtocolMapping} from 'devtools-protocol/types/protocol-mapping.js';
 import {assert} from '../util/assert.js';
 import {ProtocolError} from './Errors.js';
 import {EventEmitter} from './EventEmitter.js';
-import {Frame} from './FrameManager.js';
+import {Frame} from './Frame.js';
 import {debugError, isString} from './util.js';
 import {HTTPResponse} from './HTTPResponse.js';
