+++ src/common/HTTPResponse.ts
@@ -16,7 +16,7 @@
 import {ProtocolMapping} from 'devtools-protocol/types/protocol-mapping.js';
 
 import {EventEmitter} from './EventEmitter.js';
-import {Frame} from './FrameManager.js';
+import {Frame} from './Frame.js';
 import {HTTPRequest} from './HTTPRequest.js';
 import {SecurityDetails} from './SecurityDetails.js';
 import {Protocol} from 'devtools-protocol';
