+++ src/common/IsolatedWorld.ts
@@ -20,7 +20,8 @@ import {CDPSession} from './Connection.js';
 import {ElementHandle} from './ElementHandle.js';
 import {TimeoutError} from './Errors.js';
 import {ExecutionContext} from './ExecutionContext.js';
-import {Frame, FrameManager} from './FrameManager.js';
+import {FrameManager} from './FrameManager.js';
+import {Frame} from './Frame.js';
 import {MouseButton} from './Input.js';
 import {JSHandle} from './JSHandle.js';
 import {LifecycleWatcher, PuppeteerLifeCycleEvent} from './LifecycleWatcher.js';
