+++ test/src/utils.ts
@@ -16,7 +16,7 @@
 
 import expect from 'expect';
 import path from 'path';
-import {Frame} from '../../lib/cjs/puppeteer/common/FrameManager.js';
+import {Frame} from '../../lib/cjs/puppeteer/common/Frame.js';
 import {Page} from '../../lib/cjs/puppeteer/common/Page.js';
 import {EventEmitter} from '../../lib/cjs/puppeteer/common/EventEmitter.js';
 import {compare} from './golden-utils.js';
